requirements.txt
pandas
requests
ipykernel
tqdm
beautifulsoup4
dbt-snowflake
snowflake-connector-python[pandas]
cryptography
python-dotenv